requirements.txt
pandas
plotly-express
pyyaml
requests
streamlit
streamlit-aggrid
st_btn_select
web3


